Upgrade meow, add schism.rc and env vars
The bootstrap script only works with guile-2.2 or earlier. This
project would appear broken with guile-3.0, but in fact it appears
that compiler.ss exposes an issue with guile, as opposed to there
being a problem with schism.

The compiler itself only works with node versions up to 14. Versions
16 and 18 fail on loading the compiler image with

    CompileError: WebAssembly.instantiate(): opcode i32.rem_s is \
        not allowed in init. expressions @+5105

This opcode is required for numeric display. At present node 14 is
still LTS, but this issue needs to be fixed by the time it reaches
EOL (2023-04-30).

The cli is generated with `meow`, which had been pinned at version 5,
and that presented a problem with newer versions of node. Updating
that to 10.1.3 allowed it to be structured as an ES6 module.

Lastly, there was some cleanup in `schism.sh`, replacing an ugly "what
directory is this script in" bit of boilerplate with an env var, three
of which are defined in a new file `schism.rc`: SCHISM_GUILE,
SCHISM_NODE, and SCHISM_ROOT.